+++ src/main/java/com/rebuild/server/business/dataio/DataImporter.java
@@ -27,12 +27,12 @@ import java.util.Map;
 import org.apache.commons.lang.StringUtils;
 
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.PickListManager;
 import com.rebuild.server.helper.task.BulkTask;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.ExtRecordCreator;
 import com.rebuild.server.metadata.entityhub.DisplayType;
 import com.rebuild.server.metadata.entityhub.EasyMeta;
+import com.rebuild.server.portals.PickListManager;
 import com.rebuild.utils.JSONUtils;
 
 import cn.devezhao.commons.CalendarUtils;
