+++ src/main/java/com/rebuild/server/portals/DataListManager.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
+package com.rebuild.server.portals;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -42,7 +42,7 @@ import cn.devezhao.persist4j.engine.ID;
  * @author zhaofang123@gmail.com
  * @since 08/30/2018
  */
-public class DataListManager extends LayoutManager {
+public class DataListManager extends BaseLayoutManager {
 
 	/**
 	 * @param entity
