+++ src/main/java/com/rebuild/server/portals/ViewAddonsManager.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
+package com.rebuild.server.portals;
 
 import java.util.ArrayList;
 import java.util.Collection;
