+++ src/main/java/com/rebuild/server/portals/datalist/DataList.java
@@ -1,6 +1,6 @@
 /*
-rebuild - Building your system freely.
-Copyright (C) 2019 devezhao <zhaofang123@gmail.com>
+rebuild - Building your business-systems freely.
+Copyright (C) 2018 devezhao <zhaofang123@gmail.com>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,24 +16,34 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
-
-import org.junit.Test;
+package com.rebuild.server.portals.datalist;
 
 import com.alibaba.fastjson.JSON;
-import com.rebuild.server.TestSupport;
-import com.rebuild.server.helper.portals.DashboardManager;
-import com.rebuild.server.service.bizz.UserService;
+
+import cn.devezhao.persist4j.Entity;
 
 /**
- * @author devezhao zhaofang123@gmail.com
- * @since 2019/03/09
+ * @author Zhao Fangfang
+ * @since 1.0, 2013-6-20
  */
-public class DashboardManagerTest extends TestSupport {
-
-	@Test
-	public void testGetList() throws Exception {
-		JSON dashs = DashboardManager.getDashList(UserService.ADMIN_USER);
-		System.out.println(dashs.toJSONString());
-	}
+public interface DataList {
+	
+	/**
+	 * @return
+	 */
+	Entity getEntity();
+	
+	/**
+	 * 默认过滤条件
+	 * 
+	 * @return
+	 */
+	String getDefaultFilter();
+	
+	/**
+	 * JSON 结果集
+	 * 
+	 * @return
+	 */
+	JSON getJSONResult();
 }
