+++ src/main/java/com/rebuild/server/portals/datalist/DataWrapper.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.business.datalist;
+package com.rebuild.server.portals.datalist;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
@@ -24,9 +24,9 @@ import org.apache.commons.logging.LogFactory;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.value.FieldValueWrapper;
 import com.rebuild.server.metadata.MetadataHelper;
 import com.rebuild.server.metadata.entityhub.EasyMeta;
+import com.rebuild.server.portals.value.FieldValueWrapper;
 import com.rebuild.utils.JSONUtils;
 
 import cn.devezhao.persist4j.Entity;
