+++ src/main/java/com/rebuild/web/admin/bizz/UserControll.java
@@ -32,8 +32,8 @@ import org.springframework.web.servlet.ModelAndView;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.DataListManager;
 import com.rebuild.server.metadata.EntityHelper;
+import com.rebuild.server.portals.DataListManager;
 import com.rebuild.server.service.bizz.UserService;
 import com.rebuild.server.service.bizz.privileges.Department;
 import com.rebuild.server.service.bizz.privileges.User;
