+++ src/main/java/com/rebuild/web/admin/entityhub/FormDesignControll.java
@@ -31,8 +31,8 @@ import org.springframework.web.servlet.ModelAndView;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.FormsManager;
 import com.rebuild.server.metadata.EntityHelper;
+import com.rebuild.server.portals.FormsManager;
 import com.rebuild.web.BasePageControll;
 import com.rebuild.web.PortalsConfiguration;
