+++ src/main/java/com/rebuild/web/base/entity/GeneralModelControll.java
@@ -30,10 +30,10 @@ import org.springframework.web.servlet.ModelAndView;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
-import com.rebuild.server.helper.portals.FormsManager;
-import com.rebuild.server.helper.portals.ViewAddonsManager;
-import com.rebuild.server.helper.portals.value.DefaultValueManager;
 import com.rebuild.server.metadata.MetadataHelper;
+import com.rebuild.server.portals.FormsManager;
+import com.rebuild.server.portals.ViewAddonsManager;
+import com.rebuild.server.portals.value.DefaultValueManager;
 import com.rebuild.web.BaseEntityControll;
 
 import cn.devezhao.commons.web.ServletUtils;
