+++ src/main/java/com/rebuild/web/base/entity/RelatedListControll.java
@@ -36,9 +36,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.value.FieldValueWrapper;
 import com.rebuild.server.metadata.EntityHelper;
 import com.rebuild.server.metadata.MetadataHelper;
+import com.rebuild.server.portals.value.FieldValueWrapper;
 import com.rebuild.utils.JSONUtils;
 import com.rebuild.web.BaseControll;
