+++ src/main/java/com/rebuild/web/dashboard/DashboardControll.java
@@ -32,8 +32,8 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.Application;
-import com.rebuild.server.helper.portals.DashboardManager;
 import com.rebuild.server.metadata.EntityHelper;
+import com.rebuild.server.portals.DashboardManager;
 import com.rebuild.server.service.bizz.UserHelper;
 import com.rebuild.utils.JSONUtils;
 import com.rebuild.web.BasePageControll;
