+++ src/test/java/com/rebuild/server/portals/DashboardManagerTest.java
@@ -1,6 +1,6 @@
 /*
-rebuild - Building your business-systems freely.
-Copyright (C) 2018 devezhao <zhaofang123@gmail.com>
+rebuild - Building your system freely.
+Copyright (C) 2019 devezhao <zhaofang123@gmail.com>
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -16,13 +16,24 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
+package com.rebuild.server.portals;
+
+import org.junit.Test;
+
+import com.alibaba.fastjson.JSON;
+import com.rebuild.server.TestSupport;
+import com.rebuild.server.portals.DashboardManager;
+import com.rebuild.server.service.bizz.UserService;
 
 /**
- * 前端配置管理器。标记接口
- * 
- * @author devezhao
- * @since 01/04/2019
+ * @author devezhao zhaofang123@gmail.com
+ * @since 2019/03/09
  */
-public interface PortalsManager {
+public class DashboardManagerTest extends TestSupport {
+
+	@Test
+	public void testGetList() throws Exception {
+		JSON dashs = DashboardManager.getDashList(UserService.ADMIN_USER);
+		System.out.println(dashs.toJSONString());
+	}
 }
