+++ src/test/java/com/rebuild/server/portals/FormsManagerTest.java
@@ -16,13 +16,13 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
+package com.rebuild.server.portals;
 
 import org.junit.Test;
 
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.TestSupport;
-import com.rebuild.server.helper.portals.FormsManager;
+import com.rebuild.server.portals.FormsManager;
 import com.rebuild.server.service.bizz.UserService;
 
 /**
