+++ src/test/java/com/rebuild/server/portals/NavManagerTest.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.helper.portals;
+package com.rebuild.server.portals;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -29,7 +29,7 @@ import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.rebuild.server.TestSupport;
-import com.rebuild.server.helper.portals.NavManager;
+import com.rebuild.server.portals.NavManager;
 import com.rebuild.server.service.bizz.UserService;
 
 import cn.devezhao.commons.web.WebUtils;
