+++ src/test/java/com/rebuild/server/portals/datalist/DataListTest.java
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.business.datalist;
+package com.rebuild.server.portals.datalist;
 
 import org.junit.Before;
 import org.junit.Test;
